refactor code to satisfy sonarcloud hopefully

+++ src/iSparrow/sparrow_watcher.py
@@ -72,8 +72,6 @@ def watchertask(watcher):
         RuntimeError: When something goes wrong inside the analyzer process.
     """
 
-    print("calling watcher task")
-
     # build the recorder
     observer = Observer()
 
@@ -455,7 +453,7 @@ class SparrowWatcher:
             RuntimeError: When the watcher process is not running anymore
         """
         if self.watcher_process is not None and self.watcher_process.is_alive():
-            # print("continue the watcher process")
+            print("continue the watcher process")
             self.may_do_work.set()
         else:
             raise RuntimeError("Cannot continue watcher process, is not alive anymore.")
@@ -468,8 +466,6 @@ class SparrowWatcher:
             RuntimeError: When the watcher process is not running anymore
         """
         if self.watcher_process is not None and self.watcher_process.is_alive():
-            self.may_do_work.set()
-
             self.is_done_analyzing.wait()
 
             print("stop the watcher process")
@@ -498,37 +494,33 @@ class SparrowWatcher:
 
         for filename in self.input.iterdir():
 
-            # check that the currently checked file has been created before the last
-            # one of which we are sure it has been analyzed to make sure it does not
-            # include the one the analyzer currently uses
             condition = (
                 filename.stat().st_ctime < self.creation_time_last_analyzed.value
                 if self.is_running
                 else True
             )
 
-            if condition:
+            exists = any(
+                [
+                    (out / Path(f"results_{str(filename.stem)}.csv")).exists()
+                    for out in self.used_output_folders
+                ]
+            )
 
-                not_there = all(
-                    [
-                        (out / Path(f"results_{str(filename.stem)}.csv")).exists()
-                        is False
-                        for out in self.used_output_folders
-                    ]
-                )
+            if condition and not exists:
+                missings.append(filename)
 
-                if not_there:
-                    missings.append(filename)
-                    #  README: calling this is should not cause a race condition on the model because
-                    # the watcher_process has its own copy of the recording
-                    if self.reanalyze_on_cleanup:
-                        self.may_do_work.set()
-                        self.is_done_analyzing.clear()
-                        self.analyze(filename, recording)
+            if self.reanalyze_on_cleanup and condition and not exists:
+                recording.path = filename
+                recording.analyze()
 
-                if self.delete_recordings == "on_cleanup":
-                    # no locking needed because we do not look at files newer than the last fully analyzed one
-                    filename.unlink()
+                results = recording.detections
+
+                self.save_results(results, suffix=Path(filename).stem)
+
+            # check that the currently checked file has been created before the last
+            if self.delete_recordings in ["always", "on_cleanup"] and condition:
+                filename.unlink()
 
         with open(self.output / "missing_files.txt", "w") as missingrecord:
             for line in missings:
